requirements.txt
amqp==5.3.1
arrow==1.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==24.2.0
billiard==4.2.1
black==24.8.0
bleach==6.1.0
celery==5.4.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.1.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-cors-headers==4.4.0
django-countries==7.6.1
django-phonenumber-field==8.0.0
django-post-office==3.9.0
django-timezone-field==7.1
djangorestframework==3.15.2
drf-spectacular==0.27.2
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.4.2
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
phonenumbers==8.13.54
pillow==10.4.0
platformdirs==4.3.6
prompt_toolkit==3.0.50
psycopg2==2.9.9
PyJWT==2.9.0
pyotp==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
qrcode==8.0
redis==5.1.1
referencing==0.35.1
requests==2.32.3
rpds-py==0.20.0
six==1.16.0
sqlparse==0.5.1
tinycss2==1.2.1
tomli==2.0.2
types-python-dateutil==2.9.0.20241003
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
webencodings==0.5.1
